Organize the tests
